requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.0
ci-info==0.3.0
click==8.2.1
cohere==5.18.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
crewai==0.186.1
cryptography==46.0.1
curl_cffi==0.13.0
dataclasses-json==0.6.7
datasets==4.1.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.0
durationpy==0.10
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.116.2
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.2.10
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.11.3
invoke==2.2.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-ollama==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.28
litellm==1.74.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.2
llama-index-cli==0.5.1
llama-index-core==0.14.2
llama-index-embeddings-adapter==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-experimental==0.6.2
llama-index-finetuning==0.4.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-azure-openai==0.4.1
llama-index-llms-mistralai==0.7.1
llama-index-llms-openai==0.5.6
llama-index-postprocessor-cohere-rerank==0.5.1
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.2.2
llama-parse==0.6.54
looseversion==1.3.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.9.10
mmh3==5.2.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
multiprocess==0.70.16
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.3
oauthlib==3.3.1
ollama==0.5.4
onnxruntime==1.22.0
openai==1.108.0
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pathlib==1.0.1
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
prov==2.1.1
psutil==7.1.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.4
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.1.4
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.49
stack-data==0.6.3
starlette==0.48.0
streamlit==1.49.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tools==1.0.5
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.56.1
typer==0.17.4
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.18
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wkhtmltopdf==0.2
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
